Mostly permissions for the download lambda.
There are permissions to access the new download files lambda. There's a
couple of other bits I noticed while I was doing it which I've lumped in
together.

Jenkins didn't have access to the file format build roles. I've no idea
how this used to work but it does now.

There was a permission missing that stopped us changing the mount target
security groups, I've added it in.