Remove Jenkins backup configuration
This is being moved to the tdr-jenkins repository. Configuring it in
this repository caused a problem whenever the Jenkins EC2 instance
was redeployed, because the maintenance window is linked directly to
the EC2 _instance_ ID. This meant that when you deploy the Jenkins
EC2 instance, the backups fail because the maintenance windows tries
to run the backups on the old instance.

Moving the configuration to the tdr-jenkins repo should fix the issue
because we deploy a new EC2 instance by running the tdr-jenkins
Terraform script. This should update link between the maintenance
window and the EC2 instance whenever we deploy a new instance.