Add sns topic arns to the lambda module.
These sns topic arns are used inside the notifications lambda module to
add permissions for these sns topics to invoke the lambda function.

The idea being that we will have one sns topic each for intg, staging
and prod and these 3 will all connect to the notifications lambda in the
management account. The arns are hardcoded because terraform doesn't
have permissions to get them using a data block.