Move describe ec2 role from tdr-scripts terraform.
This is the role and policy which allows the job which checks the age of
a bastion and delete it if it is too old. The problem was that I had it
in the terraform for the bastion creation in the tdr-scripts project so
we could only check if the bastion was too old if it had already been
created which is less than ideal.
So these roles need to exist permanently so I've added them in here.