Simplify terraform roles.
I've moved most things to resource * and action *. This isn't a huge
security problem as the role still can't acesss other services we don't
use and a lot of the services, we were adding most of the permissions
anyway.
For a lot of the resources, we had policies that allowed all resources
for a given region so this is only slightly more loose.
I've left EC2 and SSM with specific actions as these services have a
huge number of actions that we're not using.
I've also specified the IAM resources explicitly so new roles can't be
created or passed without changing the policy.